ConstantExpr type update:
Types are now computed and enforced in the ConstantExpr constructor.
When a source boxed constant is passed to ConstantExpr and the computed
(restricted) type is not the same as the unboxed class type, the
constant is reboxed to the shrunken type such that the real constant
type and the computed type match.