Made LocalValueResolver context sensitive and update ExpressionEvaluator
Added empty ProductIterator because of bug in ProductIterator.

Removed TaintSet.isUnconst() and added constructors for permanently
tainted sets.

LocalValueResolver is now able to resolve the context in which a local
definition is obtained from, i.e. if a local is analysed in
ExpressionEvaluator and the definitions are Exprs from a different
method's cfg, when looking up the real value of this Expr, we search the
caller cfg instead of a statically determined one.