Created new branch with minor changes to IPA (GeomPTA)
Looks like IPA is actually the final boss, soots GeomPTA is pretty much
the best solution but implementing it here is a proper nightmare.
Storing minor changes in this branch (bad habits pushing everything
into master when we need to) until we can either come up with a proper
solution or abandon the idea and write analysis specific passes for
anything that would be otherwise IPA simulated or logic solver based.