Clean up IPA and dirty classes
Removed classes mostly pertaining to:
 - Interprocedural analysis attempts, whether external integrations or
our own native ones.
 - Invocation resolver test files and old mucky 'test system' - we will
need to write a proper framework in the future or find a way to mock and
test with junit or something similar.
 - All InvocationResolver attempts except for the vtable builder
implementation. HierarchyMethods contains the old
SimpleInvocationResolver code for hierarchy method chains which we need
to port over in the future to make use of the table structures or
similar strategy.
 - Fixed Type constant references in DumbExceptionAnalysis from
TypeUtils.<type> to use Type.getType() as the constants were removed
when caching was adding to Type.getType.

Also cleaned up boot to make no calls to anything IPA. Boot will need to
be revised in the future as well, but this is low priority because to
the other things that we need to do.