Update maven groupId's
The groupId's for all of the modules (not parent) were not set before,
so they were inheriting org.mapleir. By changing the groupId to the full
descriptor of each module (including parent) we have a better way of
tracking artefacts and modules can be split up into separate
api/implementation components more cleanly.