Restructure the project
Moved each module into a folder named the same as the module's artefact
ID. The project is now structured as follows:
root/ (git repository)
  README.md
  .gitignore
  -> dom.prod.module1/
    -> code
	-> pom.xml
  -> dom.prod.module2/
    -> code
	-> pom.xml
  -> dom.prod.parent/
    -> pom.xml

This allows for a more organised and intuitive project layout as well
as having the added benefit of not needing to make the parent folder
a project in an IDE such as Eclipse.
I chose not to nest the module folders (i.e. all modules starting with
dom. would be in the dom/ folder, and then for each subsequent
identifier, a folder would be created inside the previous identifer's
folder) as this would make the project structure more complicated and
also make it harder to rename individual modules in the future.