Add global IServiceContext to ServiceHelper
The global context can be used to share services and service factories
across different context layers. The design decision to use contexts in
the service registry API and implementation was in hindsight a bit
coarse as the context handling logic and framework could have been left
to the client to handle on the application scope. Instead, an elegant
solution is to have a context that is not able to be constrained or
targeted specifically by the application layer.