Split service-framework into API and impl modules
Split the service-framework module into it's constituent API and
implementation modules. This is a step towards more granularity and
modularity between packages and modules in the project. The dependencies
of the service-util, main and both service-framework modules have had to
be updated to reflect the changes. Otherwise, there have been no code
changes.