Move ClassServiceQuery to API module
Moved ClassServiceQuery from the service framework implementation module
to the API module, as this feature could be used throughout the project
by other client code and we don't want to have two copies of essentially
the same class in different places. We cannot push this class into the
service-util module as the latter relies on both the API and impl
modules itself.