Change BooleanProperty type to Boolean
The addition of ServiceHelper.rebasePrimitiveType means that
boolean.class can be passed as the requested type for a property instead
of Boolean.class. boolean.class is rebased to Boolean.class but this
does not work the other way around, so the type of BooleanProperty is
changed to Boolean.class to allow both boolean.class(Boolean.TYPE) and
Boolean.class to return a property of type Boolean.class.