Make DefaultInvocationResolver use a logger and fix bug
DefaultInvocationResolver could not handle resolving methods on abstract
classes (even if the method was not abstract in that class). This may be
added back in the future, as we may need to look into it.

Info and debug messages in DefaultInvocationResolver are now passed to a
logger instead of the std iostreams.