Fix bug in IRCallTracer and use logger
There was a slight bug before where if a method was not found, it would
check whether it was a constructor and if it was, would print something
different. The problem is that instead of the method name, it was
checking the owner class for the name <init>. This behaviour was removed
as the logger can be configured to filter messages instead of doing it
in the code here.