Remove FastGraphVertex.getId
Removed FastGraphVertex.getId and replaced it with a getDisplayName method that
returns a String as the display name for a given vertex. This display name may
be different to the numeric ID of a block. The reason for changing this was
because getNumericId and getId were being used in different places for the same
functionality, creating an implicit coupling in the implementation details for
the FastGraphVertex interface, if everything were to work correctly. After
going through the code to consider the getId method calls, the parts that
depended on the ID of a vertex for functionality reasons were replaced with
getNumericId and for debugging and String representations getDisplayName is now
used.

Note: no rigourous testing took place but care was taken to refactor existing
code so that it executes in essentially the exact same way.