Added hashCode and equals to AbstractFlowEdge
Previously the equality identities for flow edges was based on the
hashCode and equals methods in FastGraphEdgeImpl, which only considered
the source and destination vertices. This caused a bug for flow
graph/control flow graph edges, because it is possible to have multiple
edges from a given source block to a given destination block (e.g.
conditional statement that has the same true and false successor).
Therefore AbstractFlowEdge now considers the type of the edge as well
when checking for equality and computing the hashcode for itself.

Note: I'm not convinced that this is enough. Should FastGraphEdgeImpl
behave like this and match edges that link the same nodes, i.e. multiple
edges that are the same are considered as one link in the graph, or
should multiple edges be allowed?