Add SourcePosition framework files
SourcePosition is used to mark strings and char sequences that are
produced in the lexer or parser, so that we can mark events (e.g.
exceptions, warnings, etc) without having to compute the offsets and
information for the event every time one is created.

There are currently two implementations, one for the Tokens in Antlr4
and one for the ParseTree interface, which is by both TerminalNode
(which is used for generic lexed tokens in the grammar) and
ParserRuleContext (which represents the higher level parser rules). Each
SourcePosition is represented by line and column numbers, to represent
the start of the textual position in the parsed source file and then an
additional offset which marks for which character within the string
representation of the token the event was generated.