Add ErrorReporter and associated classes
ErrorReporter is an interface which defines the compilers exception and
warning lifecycle. The intended usage is to call a newSourcePosition
method before parsing a specific item and then releasing it afterwards
with popSourcePosition. This provides context to the ErrorReporter so
that if error or warn are invoked, there is sufficient information to
generate an appropriate event.

CompilationWarning and CompilationException both contain SourcePosition
information as well as a String message. In the case of
CompilationException, a causal Throwable can be specified. This class is
simply an extension of the checked exception mechanism in Java.

ForwardingErrorReporter is an implementation that passes on the events
that it generates to the relevant Consumers, which are able to actually
process them in whatever way they want.