Add ExpectsSourcePosition annotation
ExpectsSourcePosition is a method-bound annotation useful within the
compiler itself to indicate that before a method is called, the calling
context should've been setup in such a way, that if the callee calls the
associated ErrorReporter.error or warn methods, no internal error should
occur.

The behaviour of this annotation is difficult to verify automatically
and thus is intended to be used to help manually check the compiler
code.