Add language directive class representations
Directives are used in the language to set properties within the current
scope or in the global scope. The purpose is to provide scope specific
information that we would need to otherwise create many separate
commands for.

The value of a directive may be a literal or the name of another
declared directive, or a list of any directive value. Ideally,
expressions should be allowed to be nested inside a directive, as long
as the value can be computed at compile time.