Add basic model classes
The internal model for the compiler contains declarations that represent
the class file and the method and field attributes. These are defined
separately from the other components/modules so that the generation
information is not coupled tightly. UpdateableLocalsPool has the option
to change the state of the LocalsPool, in case the compiler wants to
make incremental changes to the code model.