Remove read(), write(Expr) and overwrite(Expr) methods from CodeUnit
These methods were not previously used anywhere in the code base. They
work in a similar fashion to their sister functions which take an index
also but instead use the current unit's last element pointer for the
child array. This means that in general it requires the user code to
track the current type of CodeUnit as well as the children of it which
makes them error prone and difficult to use.