Refactor CodeUnit.writeAt
Before writeAt would update the state of the system before realising a
potential environmental error, therefore it now fails before it changes
any data fields. Additionally it only carries out the action of changing
the children array and invoking the callback if the result of the write
would meaningfully change the system, i.e. the expression to be written
is not already attributed with this or any active unit.
Checks that were previously done in overwrite are also now done here
(bounds, capacity).

Also added javadoc for the method.