Update Pass system with context and result
Removed isQuantised from IPass.
Changed int accept(...) to PassResult accept(PathContext cxt) in IPass
and fixed subclasses to match.
Changed handling of passes in PassGroup to work with new system.

PassContext provides the same information that was previously passed by
argument to the accept method of a pass subclass. Instead of returning a
numeric delta, the pass now returns a PassResult which gives directives
to the caller as to the result and actions required by the pass after
executing.