Fix ClassTree significant performance issue due to tree grabbing all childs from parents
Short description of what happened here. The stack kept adding the parents and the childrens of the popped value from the stack.

So naturally it went like this:
ClassA -> ClassAParent -> Object -> Proceeds to add EVERY CHILD TO OBJECT AND THEIR CHILDREN TOO

The fix? Easy one. Split it into it's own while statement.